requirements.txt
annotated-types==0.7.0
anyio==4.11.0
asgiref==3.10.0
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
charset-normalizer==3.4.4
contourpy==1.3.3
cssselect2==0.8.0
cycler==0.12.1
Django==5.2.7
djangorestframework==3.16.1
fonttools==4.60.1
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.4.9
langchain==0.3.27
langchain-core==0.3.79
langchain-ollama==0.3.10
langchain-text-splitters==0.3.11
langsmith==0.4.37
matplotlib==3.10.7
narwhals==2.8.0
numpy==2.3.4
ollama==0.6.0
orjson==3.11.3
packaging==25.0
pandas==2.3.3
pillow==12.0.0
plotly==6.3.1
pycparser==2.23
pydantic==2.12.2
pydantic_core==2.41.4
pydyf==0.11.0
pyparsing==3.2.5
pyphen==0.17.2
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.44
sqlparse==0.5.3
tenacity==9.1.2
tinycss2==1.4.0
tinyhtml5==2.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
weasyprint==66.0
webencodings==0.5.1
zopfli==0.2.3.post1
zstandard==0.25.0
